updating information for handling processing of messages for the bot
